+++ RELEASE_NOTES.md
@@ -34,6 +34,7 @@ v19.1 (upcoming)
 - `Inspector` split in (and replaced by) `*Mirror` functional interfaces
 - `InspectorBinder` replaced by `AutoBinder` and `autobind()` (no args)
 - `Bindings` are created via `newBindings` and modified using the `with` methods
+- renamed method `Bindings.declareFrom` to `Bindings.declaredFrom`
 - removed class `se.jbee.inject.container.Factory` (use `Supplier`)
 - renamed class `Inject` to `Container`
 - renamed class `Typecast` to `Cast`